// apps/registry/codemods/eslint/biome/migrate-rules/types/biome.d.ts
export type PlainIndentStyle = "tab" | "space";
export type LineEnding = "lf" | "crlf" | "cr";
/**
 * Validated value for the `line_width` formatter options
 *
 * The allowed range of values is 1..=320
 */
export type LineWidth = number;
export type QuoteStyle = "double" | "single";
export type StringSet = string[];
// ...
export type Overrides = OverridePattern[];
export type VcsClientKind = "git";

/**
 * The configuration that is contained inside the file `biome.json`
 */
export interface Configuration {
  /**
   * A field for the [JSON schema](https://json-schema.org/) specification
   */
  $schema?: string | null;
  /**
   * Specific configuration for the Css language
   */
  css?: CssConfiguration | null;
  /**
   * A list of paths to other JSON files, used to extends the current configuration.
   */
  extends?: StringSet | null;
  /**
   * The configuration of the filesystem
   */
  files?: FilesConfiguration | null;
  /**
   * The configuration of the formatter
   */
  formatter?: FormatterConfiguration | null;
  /**
   * Specific configuration for the JavaScript language
   */
  javascript?: JavascriptConfiguration | null;
  /**
   * Specific configuration for the Json language
   */
  json?: JsonConfiguration | null;
  /**
   * The configuration for the linter
   */
  linter?: LinterConfiguration | null;
  /**
   * The configuration of the import sorting
   */
  organizeImports?: OrganizeImports | null;
  /**
   * A list of granular patterns that should be applied only to a sub set of files
   */
  overrides?: Overrides | null;
  /**
   * The configuration of the VCS integration
   */
  vcs?: VcsConfiguration | null;
}
/**
 * Options applied to CSS files
 */
export interface CssConfiguration {
  /**
   * Formatting options
   */
  formatter?: CssFormatter | null;
  /**
   * Parsing options
   */
  parser?: CssParser | null;
}
export interface CssFormatter {
  /**
   * Control the formatter for CSS (and its super languages) files.
   */
  enabled?: boolean | null;
  /**
   * The size of the indentation applied to CSS (and its super languages) files. Default to 2.
   */
  indentSize?: number | null;
  /**
   * The indent style applied to CSS (and its super languages) files.
   */
  indentStyle?: PlainIndentStyle | null;
  /**
   * The size of the indentation applied to CSS (and its super languages) files. Default to 2.
   */
  indentWidth?: number | null;
  /**
   * The type of line ending applied to CSS (and its super languages) files.
   */
  lineEnding?: LineEnding | null;
  /**
   * What's the max width of a line applied to CSS (and its super languages) files. Defaults to 80.
   */
  lineWidth?: LineWidth | null;
  quoteStyle?: QuoteStyle | null;
}
/**
 * Options that changes how the CSS parser behaves
 */
export interface CssParser {
  /**
   * Allow comments to appear on incorrect lines in `.css` files
   */
  allowWrongLineComments?: boolean | null;
}
/**
 * The configuration of the filesystem
 */
export interface FilesConfiguration {
  /**
   * A list of Unix shell style patterns. Biome will ignore files/folders that will match these patterns.
   */
  ignore?: StringSet | null;
  /**
   * Tells Biome to not emit diagnostics when handling files that doesn't know
   */
  ignoreUnknown?: boolean | null;
  /**
   * A list of Unix shell style patterns. Biome will handle only those files/folders that will match these patterns.
   */
  include?: StringSet | null;
  /**
   * The maximum allowed size for source code files in bytes. Files above this limit will be ignored for performance reasons. Defaults to 1 MiB
   */
  maxSize?: number | null;
}
/**
 * Generic options applied to all files
 */
export interface FormatterConfiguration {
  enabled?: boolean | null;
  /**
   * Stores whether formatting should be allowed to proceed if a given file has syntax errors
   */
  formatWithErrors?: boolean | null;
  /**
   * A list of Unix shell style patterns. The formatter will ignore files/folders that will match these patterns.
   */
  ignore?: StringSet | null;
  /**
   * A list of Unix shell style patterns. The formatter will include files/folders that will match these patterns.
   */
  include?: StringSet | null;
  /**
   * The size of the indentation, 2 by default (deprecated, use `indent-width`)
   */
  indentSize?: number | null;
  /**
   * The indent style.
   */
  indentStyle?: PlainIndentStyle | null;
  /**
   * The size of the indentation, 2 by default
   */
  indentWidth?: number | null;
  /**
   * The type of line ending.
   */
  lineEnding?: LineEnding | null;
  /**
   * What's the max width of a line. Defaults to 80.
   */
  lineWidth?: LineWidth | null;
}
/**
 * A set of options applied to the JavaScript files
 */
export interface JavascriptConfiguration {
  /**
   * Formatting options
   */
  formatter?: JavascriptFormatter | null;
  /**
   * A list of global bindings that should be ignored by the analyzers
   *
   * If defined here, they should not emit diagnostics.
   */
  globals?: StringSet | null;
  organize_imports?: JavascriptOrganizeImports | null;
  /**
   * Parsing options
   */
  parser?: JavascriptParser | null;
}
/**
 * Formatting options specific to the JavaScript files
 */
export interface JavascriptFormatter {
  /**
   * Whether to add non-necessary parentheses to arrow functions. Defaults to "always".
   */
  arrowParentheses?: ArrowParentheses | null;
  /**
   * Whether to hug the closing bracket of multiline HTML/JSX tags to the end of the last line, rather than being alone on the following line. Defaults to false.
   */
  bracketSameLine?: boolean | null;
  /**
   * Whether to insert spaces around brackets in object literals. Defaults to true.
   */
  bracketSpacing?: boolean | null;
  /**
   * Control the formatter for JavaScript (and its super languages) files.
   */
  enabled?: boolean | null;
  /**
   * The size of the indentation applied to JavaScript (and its super languages) files. Default to 2.
   */
  indentSize?: number | null;
  /**
   * The indent style applied to JavaScript (and its super languages) files.
   */
  indentStyle?: PlainIndentStyle | null;
  /**
   * The size of the indentation applied to JavaScript (and its super languages) files. Default to 2.
   */
  indentWidth?: number | null;
  /**
   * The type of quotes used in JSX. Defaults to double.
   */
  jsxQuoteStyle?: QuoteStyle | null;
  /**
   * The type of line ending applied to JavaScript (and its super languages) files.
   */
  lineEnding?: LineEnding | null;
  /**
   * What's the max width of a line applied to JavaScript (and its super languages) files. Defaults to 80.
   */
  lineWidth?: LineWidth | null;
  /**
   * When properties in objects are quoted. Defaults to asNeeded.
   */
  quoteProperties?: QuoteProperties | null;
  /**
   * The type of quotes used in JavaScript code. Defaults to double.
   */
  quoteStyle?: QuoteStyle | null;
  /**
   * Whether the formatter prints semicolons for all statements or only in for statements where it is necessary because of ASI.
   */
  semicolons?: Semicolons | null;
  /**
   * Print trailing commas wherever possible in multi-line comma-separated syntactic structures. Defaults to "all".
   */
  trailingComma?: TrailingComma | null;
}
export interface JavascriptOrganizeImports {}
/**
 * Options that changes how the JavaScript parser behaves
 */
export interface JavascriptParser {
  /**
   * It enables the experimental and unsafe parsing of parameter decorators
   *
   * These decorators belong to an old proposal, and they are subject to change.
   */
  unsafeParameterDecoratorsEnabled?: boolean | null;
}
/**
 * Options applied to JSON files
 */
export interface JsonConfiguration {
  /**
   * Formatting options
   */
  formatter?: JsonFormatter | null;
  /**
   * Parsing options
   */
  parser?: JsonParser | null;
}
export interface JsonFormatter {
  /**
   * Control the formatter for JSON (and its super languages) files.
   */
  enabled?: boolean | null;
  /**
   * The size of the indentation applied to JSON (and its super languages) files. Default to 2.
   */
  indentSize?: number | null;
  /**
   * The indent style applied to JSON (and its super languages) files.
   */
  indentStyle?: PlainIndentStyle | null;
  /**
   * The size of the indentation applied to JSON (and its super languages) files. Default to 2.
   */
  indentWidth?: number | null;
  /**
   * The type of line ending applied to JSON (and its super languages) files.
   */
  lineEnding?: LineEnding | null;
  /**
   * What's the max width of a line applied to JSON (and its super languages) files. Defaults to 80.
   */
  lineWidth?: LineWidth | null;
}
/**
 * Options that changes how the JSON parser behaves
 */
export interface JsonParser {
  /**
   * Allow parsing comments in `.json` files
   */
  allowComments?: boolean | null;
  /**
   * Allow parsing trailing commas in `.json` files
   */
  allowTrailingCommas?: boolean | null;
}
export interface LinterConfiguration {
  /**
   * if `false`, it disables the feature and the linter won't be executed. `true` by default
   */
  enabled?: boolean | null;
  /**
   * A list of Unix shell style patterns. The formatter will ignore files/folders that will match these patterns.
   */
  ignore?: StringSet | null;
  /**
   * A list of Unix shell style patterns. The formatter will include files/folders that will match these patterns.
   */
  include?: StringSet | null;
  /**
   * List of rules
   */
  rules?: Rules | null;
}
export interface Rules {
  a11y?: A11Y | null;
  /**
   * It enables ALL rules. The rules that belong to `nursery` won't be enabled.
   */
  all?: boolean | null;
  complexity?: Complexity | null;
  correctness?: Correctness | null;
  nursery?: Nursery | null;
  performance?: Performance | null;
  /**
   * It enables the lint rules recommended by Biome. `true` by default.
   */
  recommended?: boolean | null;
  security?: Security | null;
  style?: Style | null;
  suspicious?: Suspicious | null;
}
// ...
export interface OrganizeImports {
  /**
   * Enables the organization of imports
   */
  enabled?: boolean | null;
  /**
   * A list of Unix shell style patterns. The formatter will ignore files/folders that will match these patterns.
   */
  ignore?: StringSet | null;
  /**
   * A list of Unix shell style patterns. The formatter will include files/folders that will match these patterns.
   */
  include?: StringSet | null;
}
export interface OverridePattern {
  /**
   * Specific configuration for the Css language
   */
  css?: CssConfiguration | null;
  /**
   * Specific configuration for the Json language
   */
  formatter?: OverrideFormatterConfiguration | null;
  /**
   * A list of Unix shell style patterns. The formatter will ignore files/folders that will match these patterns.
   */
  ignore?: StringSet | null;
  /**
   * A list of Unix shell style patterns. The formatter will include files/folders that will match these patterns.
   */
  include?: StringSet | null;
  /**
   * Specific configuration for the JavaScript language
   */
  javascript?: JavascriptConfiguration | null;
  /**
   * Specific configuration for the Json language
   */
  json?: JsonConfiguration | null;
  /**
   * Specific configuration for the Json language
   */
  linter?: OverrideLinterConfiguration | null;
  /**
   * Specific configuration for the Json language
   */
  organizeImports?: OverrideOrganizeImportsConfiguration | null;
}
// ...
export interface VcsConfiguration {
  /**
   * The kind of client.
   */
  clientKind?: VcsClientKind | null;
  /**
   * The main branch of the project
   */
  defaultBranch?: string | null;
  /**
   * Whether Biome should integrate itself with the VCS client
   */
  enabled?: boolean | null;
  /**
   * The folder where Biome should check for VCS files. By default, Biome will use the same folder where `biome.json` was found.
   *
   * If Biome can't find the configuration, it will attempt to use the current working directory. If no current working directory can't be found, Biome won't use the VCS integration, and a diagnostic will be emitted
   */
  root?: string | null;
  /**
   * Whether Biome should use the VCS ignore file. When [true], Biome will ignore the files specified in the ignore file.
   */
  useIgnoreFile?: boolean | null;
}