// apps/registry/codemods/eslint/biome/migrate-rules/types/eslint.d.ts
export type Plugins = string[];
/**
 * ESLint comes with a large number of rules. You can modify which rules your project uses either using configuration comments or configuration files.
 */
export type Rules = PossibleErrors &
  BestPractices &
  StrictMode &
  Variables &
  NodeAndCommonJs &
  StylisticIssues &
  EcmaScript6 &
  Legacy &
  HttpsJsonSchemastoreOrgPartialEslintPluginsJson;
export type Rule = number | ("off" | "warn" | "error") | unknown[];
/**
 * Allows to override configuration for files and folders, specified by glob patterns
 */
export type Overrides = {
  /**
   * Glob pattern for files to apply 'overrides' configuration, relative to the directory of the config file
   */
  files: string | [string, ...string[]];
  /**
   * If you want to extend a specific configuration file, you can use the extends property and specify the path to the file. The path can be either relative or absolute.
   */
  extends?: string | string[];
  /**
   * If a file matches any of the 'excludedFiles' glob patterns, the 'overrides' configuration won't apply
   */
  excludedFiles?: string | string[];
  ecmaFeatures?: EcmaFeatures;
  env?: Env;
  globals?: Globals;
  parser?: string;
  parserOptions?: ParserOptions;
  plugins?: Plugins;
  /**
   * To specify a processor, specify the plugin name and processor name joined by a forward slash
   */
  processor?: string;
  rules?: Rules;
  settings?: Settings;
  overrides?: Overrides;
}[];

export interface JSONSchemaForESLintConfigurationFiles {
  ecmaFeatures?: EcmaFeatures;
  env?: Env;
  /**
   * If you want to extend a specific configuration file, you can use the extends property and specify the path to the file. The path can be either relative or absolute.
   */
  extends?: string | string[];
  globals?: Globals;
  /**
   * Prevent comments from changing config or rules
   */
  noInlineConfig?: boolean;
  parser?: string;
  parserOptions?: ParserOptions;
  plugins?: Plugins;
  /**
   * By default, ESLint will look for configuration files in all parent folders up to the root directory. This can be useful if you want all of your projects to follow a certain convention, but can sometimes lead to unexpected results. To limit ESLint to a specific project, set this to `true` in a configuration in the root of your project.
   */
  root?: boolean;
  /**
   * Tell ESLint to ignore specific files and directories. Each value uses the same pattern as the `.eslintignore` file.
   */
  ignorePatterns?: string | string[];
  rules?: Rules;
  settings?: Settings;
  overrides?: Overrides;
  [k: string]: unknown;
}
/**
 * By default, ESLint supports only ECMAScript 5 syntax. You can override that setting to enable support for ECMAScript 6 as well as JSX by using configuration settings.
 */
export interface EcmaFeatures {
  arrowFunctions?: boolean;
  binaryLiterals?: boolean;
  blockBindings?: boolean;
  classes?: boolean;
  defaultParams?: boolean;
  destructuring?: boolean;
  /**
   * Enables support for the experimental object rest/spread properties (IMPORTANT: This is an experimental feature that may change significantly in the future. It's recommended that you do not write rules relying on this functionality unless you are willing to incur maintenance cost when it changes.)
   */
  experimentalObjectRestSpread?: boolean;
  forOf?: boolean;
  generators?: boolean;
  /**
   * allow return statements in the global scope
   */
  globalReturn?: boolean;
  /**
   * enable global strict mode (if ecmaVersion is 5 or greater)
   */
  impliedStrict?: boolean;
  /**
   * enable JSX
   */
  jsx?: boolean;
  modules?: boolean;
  objectLiteralComputedProperties?: boolean;
  objectLiteralDuplicateProperties?: boolean;
  objectLiteralShorthandMethods?: boolean;
  objectLiteralShorthandProperties?: boolean;
  octalLiterals?: boolean;
  regexUFlag?: boolean;
  regexYFlag?: boolean;
  restParams?: boolean;
  spread?: boolean;
  superInFunctions?: boolean;
  templateStrings?: boolean;
  unicodeCodePointEscapes?: boolean;
  [k: string]: unknown;
}
/**
 * An environment defines global variables that are predefined.
 */
export interface Env {
  /**
   * defines require() and define() as global variables as per the amd spec
   */
  amd?: boolean;
  /**
   * AppleScript global variables
   */
  applescript?: boolean;
  /**
   * Atom test helper globals
   */
  atomtest?: boolean;
  /**
   * browser global variables
   */
  browser?: boolean;
  /**
   * CommonJS global variables and CommonJS scoping (use this for browser-only code that uses Browserify/WebPack)
   */
  commonjs?: boolean;
  /**
   * Globals common to both Node and Browser
   */
  "shared-node-browser"?: boolean;
  /**
   * Ember test helper globals
   */
  embertest?: boolean;
  /**
   * enable all ECMAScript 6 features except for modules
   */
  es6?: boolean;
  /**
   * GreaseMonkey globals
   */
  greasemonkey?: boolean;
  /**
   * adds all of the Jasmine testing global variables for version 1.3 and 2.0
   */
  jasmine?: boolean;
  /**
   * Jest global variables
   */
  jest?: boolean;
  /**
   * jQuery global variables
   */
  jquery?: boolean;
  /**
   * Meteor global variables
   */
  meteor?: boolean;
  /**
   * adds all of the Mocha test global variables
   */
  mocha?: boolean;
  /**
   * MongoDB global variables
   */
  mongo?: boolean;
  /**
   * Java 8 Nashorn global variables
   */
  nashorn?: boolean;
  /**
   * Node.js global variables and Node.js scoping
   */
  node?: boolean;
  /**
   * PhantomJS global variables
   */
  phantomjs?: boolean;
  /**
   * Prototype.js global variables
   */
  prototypejs?: boolean;
  /**
   * Protractor global variables
   */
  protractor?: boolean;
  /**
   * QUnit global variables
   */
  qunit?: boolean;
  /**
   * Service Worker global variables
   */
  serviceworker?: boolean;
  /**
   * ShellJS global variables
   */
  shelljs?: boolean;
  /**
   * WebExtensions globals
   */
  webextensions?: boolean;
  /**
   * web workers global variables
   */
  worker?: boolean;
  [k: string]: unknown;
}
/**
 * Set each global variable name equal to true to allow the variable to be overwritten or false to disallow overwriting.
 */
export interface Globals {
  [k: string]: ("readonly" | "writable" | "off") | boolean;
}
/**
 * The JavaScript language options to be supported
 */
export interface ParserOptions {
  ecmaFeatures?: EcmaFeatures;
  /**
   * Set to 3, 5, 6, 7, 8, 9, 10, 11 (default), 12, 13, 14 or "latest" to specify the version of ECMAScript syntax you want to use. You can also set to 2015 (same as 6), 2016 (same as 7), 2017 (same as 8), 2018 (same as 9), 2019 (same as 10), 2020 (same as 11), 2021 (same as 12), 2022 (same as 13) or 2023 (same as 14) to use the year-based naming. "latest" always enables the latest supported ECMAScript version.
   */
  ecmaVersion?:
    | 3
    | 5
    | 6
    | 2015
    | 7
    | 2016
    | 8
    | 2017
    | 9
    | 2018
    | 10
    | 2019
    | 11
    | 2020
    | 12
    | 2021
    | 13
    | 2022
    | 14
    | 2023
    | "latest";
  /**
   * set to "script" (default), "commonjs", or "module" if your code is in ECMAScript modules
   */
  sourceType?: "script" | "module" | "commonjs";
  [k: string]: unknown;
}
export interface PossibleErrors {
  /**
   * Require or disallow trailing commas
   */
  "comma-dangle"?: number | ("off" | "warn" | "error") | unknown[];
  /**
   * Enforce "for" loop update clause moving the counter in the right direction
   */
  "for-direction"?: number | ("off" | "warn" | "error") | unknown[];
  /**
   * Enforce return statements in getters
   */
  "getter-return"?: number | ("off" | "warn" | "error") | unknown[];
  /**
   * Disallow await inside of loops
   */
  "no-await-in-loop"?: number | ("off" | "warn" | "error") | unknown[];
  /**
   * Disallow comparing against -0
   */
  "no-compare-neg-zero"?: number | ("off" | "warn" | "error") | unknown[];
  /**
   * Disallow assignment operators in conditional expressions
   */
  "no-cond-assign"?: number | ("off" | "warn" | "error") | unknown[];
  /**
   * Disallow the use of console
   */
  "no-console"?: number | ("off" | "warn" | "error") | unknown[];
  /**
   * Disallow constant expressions in conditions
   */
  "no-constant-condition"?: number | ("off" | "warn" | "error") | unknown[];
  /**
   * Disallow control characters in regular expressions
   */
  "no-control-regex"?: number | ("off" | "warn" | "error") | unknown[];
  /**
   * Disallow the use of debugger
   */
  "no-debugger"?: number | ("off" | "warn" | "error") | unknown[];
  /**
   * Disallow duplicate arguments in function definitions
   */
  "no-dupe-args"?: number | ("off" | "warn" | "error") | unknown[];
  /**
   * Disallow duplicate keys in object literals
   */
  "no-dupe-keys"?: number | ("off" | "warn" | "error") | unknown[];
  /**
   * Disallow duplicate case labels
   */
  "no-duplicate-case"?: number | ("off" | "warn" | "error") | unknown[];
  /**
   * Disallow empty block statements
   */
  "no-empty"?: number | ("off" | "warn" | "error") | unknown[];
  /**
   * Disallow empty character classes in regular expressions
   */
  "no-empty-character-class"?: number | ("off" | "warn" | "error") | unknown[];
  /**
   * Disallow reassigning exceptions in catch clauses
   */
  "no-ex-assign"?: number | ("off" | "warn" | "error") | unknown[];
  /**
   * Disallow unnecessary boolean casts
   */
  "no-extra-boolean-cast"?: number | ("off" | "warn" | "error") | unknown[];
  /**
   * Disallow unnecessary parentheses
   */
  "no-extra-parens"?: number | ("off" | "warn" | "error") | unknown[];
  /**
   * Disallow unnecessary semicolons
   */
  "no-extra-semi"?: number | ("off" | "warn" | "error") | unknown[];
  /**
   * Disallow reassigning function declarations
   */
  "no-func-assign"?: number | ("off" | "warn" | "error") | unknown[];
  /**
   * Disallow function or var declarations in nested blocks
   */
  "no-inner-declarations"?: number | ("off" | "warn" | "error") | unknown[];
  /**
   * Disallow invalid regular expression strings in RegExp constructors
   */
  "no-invalid-regexp"?: number | ("off" | "warn" | "error") | unknown[];
  /**
   * Disallow irregular whitespace outside of strings and comments
   */
  "no-irregular-whitespace"?: number | ("off" | "warn" | "error") | unknown[];
  /**
   * Disallow negating the left operand in in expressions (deprecated)
   */
  "no-negated-in-lhs"?: number | ("off" | "warn" | "error") | unknown[];
  /**
   * Disallow calling global object properties as functions
   */
  "no-obj-calls"?: number | ("off" | "warn" | "error") | unknown[];
  /**
   * Disallow calling some Object.prototype methods directly on objects
   */
  "no-prototype-builtins"?: number | ("off" | "warn" | "error") | unknown[];
  /**
   * Disallow multiple spaces in regular expressions
   */
  "no-regex-spaces"?: number | ("off" | "warn" | "error") | unknown[];
  /**
   * Disallow sparse arrays
   */
  "no-sparse-arrays"?: number | ("off" | "warn" | "error") | unknown[];
  /**
   * Disallow template literal placeholder syntax in regular strings
   */
  "no-template-curly-in-string"?: number | ("off" | "warn" | "error") | unknown[];
  /**
   * Disallow confusing multiline expressions
   */
  "no-unexpected-multiline"?: number | ("off" | "warn" | "error") | unknown[];
  /**
   * Disallow unreachable code after return, throw, continue, and break statements
   */
  "no-unreachable"?: number | ("off" | "warn" | "error") | unknown[];
  /**
   * Disallow control flow statements in finally blocks
   */
  "no-unsafe-finally"?: number | ("off" | "warn" | "error") | unknown[];
  /**
   * Disallow negating the left operand of relational operators
   */
  "no-unsafe-negation"?: number | ("off" | "warn" | "error") | unknown[];
  /**
   * Require calls to isNaN() when checking for NaN
   */
  "use-isnan"?: number | ("off" | "warn" | "error") | unknown[];
  /**
   * Enforce valid JSDoc comments
   */
  "valid-jsdoc"?: number | ("off" | "warn" | "error") | unknown[];
  /**
   * Enforce comparing typeof expressions against valid strings
   */
  "valid-typeof"?: number | ("off" | "warn" | "error") | unknown[];
  [k: string]: unknown;
}
// ...
export interface StrictMode {
  /**
   * require or disallow strict mode directives
   */
  strict?: number | ("off" | "warn" | "error") | unknown[];
  [k: string]: unknown;
}
// ...
export interface Legacy {
  "max-depth"?: Rule;
  "max-len"?: Rule;
  "max-params"?: Rule;
  "max-statements"?: Rule;
  "no-bitwise"?: Rule;
  "no-plusplus"?: Rule;
  [k: string]: unknown;
}
// ...
export interface Settings {
  [k: string]: unknown;
}